move admin tests into unit; fix
