+++ src/admin/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod media;
 pub(crate) mod query;
 pub(crate) mod room;
 pub(crate) mod server;
+mod tests;
 pub(crate) mod user;
 pub(crate) mod utils;
 
@@ -52,31 +53,4 @@ pub async fn fini() {
 		.write()
 		.expect("locked for writing")
 		.take();
-}
-
-#[cfg(test)]
-mod test {
-	use clap::Parser;
-
-	use crate::handler::AdminCommand;
-
-	#[test]
-	fn get_help_short() { get_help_inner("-h"); }
-
-	#[test]
-	fn get_help_long() { get_help_inner("--help"); }
-
-	#[test]
-	fn get_help_subcommand() { get_help_inner("help"); }
-
-	fn get_help_inner(input: &str) {
-		let error = AdminCommand::try_parse_from(["argv[0] doesn't matter", input])
-			.unwrap_err()
-			.to_string();
-
-		// Search for a handful of keywords that suggest the help printed properly
-		assert!(error.contains("Usage:"));
-		assert!(error.contains("Commands:"));
-		assert!(error.contains("Options:"));
-	}
 }
